fix: switch chain button without modal
